Census layers: move all logic into hook
