+++ src/components/map/config.census.ts
@@ -5,6 +5,9 @@ const tractsSrcId: LocalTypes.CensusScope = 'tract'
 const pumaSrcID: LocalTypes.CensusScope = 'puma'
 const minzoom = 8
 
+// Tell MB not to use default `id` as unique ID
+export const CENSUS_PROMOTE_ID_FIELD = 'GEOID'
+
 export const pumaLyrSrc = {
   source: pumaSrcID,
   'source-layer': 'NYC_PUMA2017_5yr_langHome-0tfhwp',
